+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Flutter: share a program as text file.
 - Flutter: replace the manual as markdown with an html version.
+- Flutter: fix "Failed to start audio playback".
 
 ### 1.0_76